Major features in this patch are: Merge of Stas' latest source into client.  - Lots of double buffering code  - Cleanup of protoyping functions and proper variable initialization  - Addition of working RSA key generator mkkey.exe (this necessitates another DLL in the source)  - Updated compile instructions, and a new document on how to make a RSA key  - Working version of winkey with BCC compiler  - Bug fixes as per listed in his change log Cleanup of changes list to remove bug fix/stuff only coders need to know. Removal of buildexe script - Stas rewrote build to make this obsolete Addition of HR bitmap set (including a few placeholder bitmaps until art is done) Metablock patch accepted (not working proper yet though due to retrieving login issue) Probably a few other things I forgot about!

+++ src/cowmain.c
@@ -955,7 +955,7 @@ cowmain (char *server,
             if (getteam[i] == me->p_mapchars[0])
                 break;
 
-        me->p_team = i;
+        me->p_team = (short) i;
 
         if (me->p_damage > me->p_ship.s_maxdamage)
         {
